backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
django-cors-headers==4.3.1
django-filter==23.3
reportlab==4.0.4
drf-spectacular==0.26.2